requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.24.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
gtts==2.4.0
pyttsx3==2.90
reportlab==4.0.7
python-json-logger==2.0.7
